refactor(split): move peripheral service out of Bluetooth directory
