+++ app/src/pointing/input_split.c
@@ -59,8 +59,7 @@ int zmk_input_split_report_peripheral_event(uint8_t reg, uint8_t type, uint16_t 
             zmk_input_processor_handle_event(processors_##n[i].dev, evt, processors_##n[i].param1, \
                                              processors_##n[i].param2, NULL);                      \
         }                                                                                          \
-        zmk_split_bt_report_input(DT_INST_REG_ADDR(n), evt->type, evt->code, evt->value,           \
-                                  evt->sync);                                                      \
+        zmk_split_report_input(DT_INST_REG_ADDR(n), evt->type, evt->code, evt->value, evt->sync);  \
     }                                                                                              \
     INPUT_CALLBACK_DEFINE(DEVICE_DT_GET(DT_INST_PHANDLE(n, device)), split_input_handler_##n);
