+++ app/src/split/CMakeLists.txt
@@ -3,6 +3,7 @@
 
 if (CONFIG_ZMK_SPLIT_BLE AND (NOT CONFIG_ZMK_SPLIT_ROLE_CENTRAL))
   target_sources(app PRIVATE listener.c)
+  target_sources(app PRIVATE service.c)
 endif()
 
 if (CONFIG_ZMK_SPLIT_BLE)
