+++ app/src/split/listener.c
@@ -7,7 +7,7 @@
 #include <zephyr/device.h>
 #include <zephyr/logging/log.h>
 
-#include <zmk/split/bluetooth/service.h>
+#include <zmk/split/service.h>
 
 LOG_MODULE_DECLARE(zmk, CONFIG_ZMK_LOG_LEVEL);
 
@@ -23,17 +23,17 @@ int split_listener(const zmk_event_t *eh) {
     const struct zmk_position_state_changed *pos_ev;
     if ((pos_ev = as_zmk_position_state_changed(eh)) != NULL) {
         if (pos_ev->state) {
-            return zmk_split_bt_position_pressed(pos_ev->position);
+            return zmk_split_position_pressed(pos_ev->position);
         } else {
-            return zmk_split_bt_position_released(pos_ev->position);
+            return zmk_split_position_released(pos_ev->position);
         }
     }
 
 #if ZMK_KEYMAP_HAS_SENSORS
     const struct zmk_sensor_event *sensor_ev;
     if ((sensor_ev = as_zmk_sensor_event(eh)) != NULL) {
-        return zmk_split_bt_sensor_triggered(sensor_ev->sensor_index, sensor_ev->channel_data,
-                                             sensor_ev->channel_data_size);
+        return zmk_split_sensor_triggered(sensor_ev->sensor_index, sensor_ev->channel_data,
+                                          sensor_ev->channel_data_size);
     }
 #endif /* ZMK_KEYMAP_HAS_SENSORS */
     return ZMK_EV_EVENT_BUBBLE;
